Landing page successfully displays number of existing number of users within the database
